try set gasPrice

+++ lastwill/contracts/submodels/ico.py
@@ -186,7 +186,7 @@ class AbstractContractDetailsICO(CommonDetails):
                 contract_data=binascii.hexlify(tr.encode_function_call(
                     'transferOwnership', [self.eth_contract_crowdsale.address]
                 )).decode(),
-                gas_price=41 * 10 ** 9 * 1.2
+                gas_price=int(41 * 10 ** 9 * 1.2)
             )
             self.eth_contract_token.tx_hash = eth_int.eth_sendRawTransaction(
                 '0x' + signed_data
